fix: remove production-specific API documentation and routing configurations

+++ core/lib/Thelia/Form/BruteforceForm.php
@@ -30,11 +30,11 @@ abstract class BruteforceForm extends FirewallForm
 
     public function getConfigTime(): int
     {
-        return ConfigQuery::read('form_firewall_bruteforce_time_to_wait', static::DEFAULT_TIME_TO_WAIT);
+        return (int) ConfigQuery::read('form_firewall_bruteforce_time_to_wait', static::DEFAULT_TIME_TO_WAIT);
     }
 
     public function getConfigAttempts(): int
     {
-        return ConfigQuery::read('form_firewall_bruteforce_attempts', static::DEFAULT_ATTEMPTS);
+        return (int) ConfigQuery::read('form_firewall_bruteforce_attempts', static::DEFAULT_ATTEMPTS);
     }
 }
